% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtroc.R
\name{CDFEst}
\alias{CDFEst}
\title{estimate the distribution function in the diseased (case) and non-disease (control) population}
\usage{
CDFEst(Xgroup, Mvec, pi0, pi1, K, lambda)
}
\description{
estimate the distribution function in the diseased (case) and non-disease (control) population
}
