% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtroc.R
\name{gtroc}
\alias{gtroc}
\title{Nonparametric Estimation of Distributions and Diagnostic Accuracy Based on Group-Tested Results with Differential Misclassification}
\usage{
gtroc(pi0 = 0.9, pi1 = 0.9, lambda = 0.02, tempDat, nodeNum = 100, ...)
}
\description{
This package concerns the problem of estimating a continuous distribution in a diseased or
nondiseased population when only group-based test results on the disease status are available. The problem
 is challenging in that individual disease statuses are not observed and testing results are often subject
 to misclassification, with further complication that the misclassification may be differential as the group
 size and the number of the diseased individuals in the group vary. We propose a method to construct
 nonparametric estimation of the distribution and obtain its asymptotic properties. The performance of the
 distribution estimator is evaluated under various design considerations concerning group sizes and
 classiffication errors. The method is exemplified with data from the National Health and Nutrition
 Examination Survey study to estimate the distribution and diagnostic accuracy of C-reactive protein in
 blood samples in predicting chlamydia incidence.

the ROC curve estimate  based on the estimated isotonic CDF
}
