% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtroc.R
\name{Loglik_der1}
\alias{Loglik_der1}
\title{the first derivative of the loglikelihood function for the prevalence}
\usage{
Loglik_der1(n0, n1, p, K, pi0, pi1, lambda)
}
\description{
the first derivative of the loglikelihood function for the prevalence
}
